Fix type error bug.

+++ ExcitedMap/src/main/java/com/excitedmap/dao/SpotMapperImpl.java
@@ -24,7 +24,7 @@ public interface SpotMapperImpl extends SpotMapper {
 
 	List<Spot> selectBySpotNameKeyword(String keyword);
 	
-	Integer selectAverageReviewRatingForSpot(int spotId);
+	Double selectAverageReviewRatingForSpot(int spotId);
 	
 	List<Map<String, Integer>> selectReviewRatingCountForSpot(int spotId);
 }
